:recycle: Move log functions out of print util
